python/requirements.txt
tensorflow
keras
pillow
opencv-python
matplotlib
flask
flask_cors
gunicorn